Fix OAuth implementation with proper callback endpoint
- Created dedicated /api/auth/callback endpoint for OAuth callback
- Updated /api/auth/github to only handle OAuth initiation
- Removed auth-button component and implemented inline auth
- Removed header, integrated OAuth into existing button
- Button now shows "Sign in with GitHub" when not authenticated
- Button shows "Add to GitHub repository" when authenticated
- Follows GitHub OAuth documentation with proper callback URL
