+++ app/api/auth/github/route.ts
@@ -1,87 +1,14 @@
 import { NextRequest, NextResponse } from 'next/server';
 
 const GITHUB_CLIENT_ID = process.env.GITHUB_CLIENT_ID || 'Ov23liJ0xVB8vtPP1apH';
-const GITHUB_CLIENT_SECRET = process.env.GITHUB_CLIENT_SECRET;
-
-if (!GITHUB_CLIENT_SECRET) {
-  throw new Error('GITHUB_CLIENT_SECRET is required');
-}
 
 export async function GET(request: NextRequest) {
-  const { searchParams } = new URL(request.url);
-  const code = searchParams.get('code');
-  const state = searchParams.get('state');
+  // Start OAuth flow - redirect to GitHub authorization
+  const authUrl = new URL('https://github.com/login/oauth/authorize');
+  authUrl.searchParams.set('client_id', GITHUB_CLIENT_ID);
+  authUrl.searchParams.set('redirect_uri', `${new URL(request.url).origin}/api/auth/callback`);
+  authUrl.searchParams.set('scope', 'user:email');
+  authUrl.searchParams.set('state', Math.random().toString(36).substring(7));
 
-  if (!code) {
-    // Start OAuth flow
-    const authUrl = new URL('https://github.com/login/oauth/authorize');
-    authUrl.searchParams.set('client_id', GITHUB_CLIENT_ID);
-    authUrl.searchParams.set('redirect_uri', `${new URL(request.url).origin}/api/auth/github`);
-    authUrl.searchParams.set('scope', 'user:email');
-    authUrl.searchParams.set('state', Math.random().toString(36).substring(7));
-
-    return NextResponse.redirect(authUrl.toString());
-  }
-
-  try {
-    // Exchange code for access token
-    const tokenResponse = await fetch('https://github.com/login/oauth/access_token', {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        client_id: GITHUB_CLIENT_ID,
-        client_secret: GITHUB_CLIENT_SECRET,
-        code,
-      }),
-    });
-
-    const tokenData = await tokenResponse.json();
-
-    if (tokenData.error) {
-      return NextResponse.json({ error: tokenData.error }, { status: 400 });
-    }
-
-    // Get user info
-    const userResponse = await fetch('https://api.github.com/user', {
-      headers: {
-        'Authorization': `Bearer ${tokenData.access_token}`,
-        'Accept': 'application/vnd.github.v3+json',
-      },
-    });
-
-    const userData = await userResponse.json();
-
-    // Here you would typically store the user data and access token in your database
-    // For now, we'll just redirect to the home page with user info
-    const response = NextResponse.redirect(new URL('/', request.url));
-    
-    // Set secure cookies
-    response.cookies.set('github_access_token', tokenData.access_token, {
-      httpOnly: true,
-      secure: process.env.NODE_ENV === 'production',
-      sameSite: 'lax',
-      maxAge: 60 * 60 * 24 * 30, // 30 days
-    });
-
-    response.cookies.set('github_user', JSON.stringify({
-      id: userData.id,
-      login: userData.login,
-      name: userData.name,
-      email: userData.email,
-      avatar_url: userData.avatar_url,
-    }), {
-      httpOnly: true,
-      secure: process.env.NODE_ENV === 'production',
-      sameSite: 'lax',
-      maxAge: 60 * 60 * 24 * 30, // 30 days
-    });
-
-    return response;
-  } catch (error) {
-    console.error('GitHub OAuth error:', error);
-    return NextResponse.json({ error: 'Authentication failed' }, { status: 500 });
-  }
+  return NextResponse.redirect(authUrl.toString());
 }
